% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArboMAP_package_form.R
\name{ArboMAP}
\alias{ArboMAP}
\title{ArboMAP (main external function call)}
\usage{
ArboMAP(humandatafile, mosqfile, districtshapefile, stratafile,
  weatherpathstr, weathersummaryfile, maxobservedhumandate, weekinquestion,
  var1name, var2name, compyear1, compyear2, results.path,
  original.metrics = 0)
}
\arguments{
\item{humandatafile}{Data on human cases, by date of onset. For models that do
not consider date of onset, these data will be recompiled. This can be given as the name of a text file, or can be entered as an R data object.}

\item{mosqfile}{Data on district of collection, collection date,
west nile virus pool result, size of the mosquito pool, and species. This can be the name of a text file or as an R data object.}

\item{districtshapefile}{A shapefile containing the districts in polygon format.}

\item{stratafile}{A text file with data on how to stratifiy the district polygons}

\item{weatherpathstr}{A directory containing the weather data to be aggregated.
The weather data can contain multiple files, and these can overlap in date range,
as long as they all have the same columns.
The weather data files in this folder should have a district column, a doy (day of year) column
a year column, and then an entry for each independent variable of interest}

\item{weathersummaryfile}{If this file is in weatherpathstr, it will be ignored.
Note that no consolidated summary file is currently being created.}

\item{maxobservedhumandate}{The last date for which human data should be included. This should NOT include the year for which forecasts are made.}

\item{weekinquestion}{The week to be used in the analysis}

\item{var1name}{The first variable to be used in the model. These can be identified using the model selection code.}

\item{var2name}{The second variable to be used in the model.}

\item{compyear1}{The first year to use for comparison to the current year}

\item{compyear2}{The second comparison year}

\item{results.path}{The path for the model results}

\item{original.metrics}{1 runs the code to correspond to the original version of ArboMAP. 0 makes departures from the code for the calculation of number of human cases predicted in the year}
}
\description{
ArboMAP is a set of software to be used in the RStudio envionment to model and predict vector-borne diseases, especially arboviruses transmitted by mosquitoes. ArboMAP was developed by the EcoGRAPH research group at the University of Oklahoma with support from NASA. We are happy to answer your questions, and we would appreciate feedback on how these tools could be improved. Please contact justinkdavis@ou.edu for technical issues about the original code, or mcwimberly@ou.edu for questions about the arbovirus modeling project. For issues related to the package/function modifications, please contact akeyel@albany.edu.
}
