% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArboMAP_package_form.R
\name{simplifynames}
\alias{simplifynames}
\title{Simplify Names (ArboMAP)}
\usage{
simplifynames(priornames = NULL)
}
\arguments{
\item{priornames}{A vector of names to be simplified}
}
\description{
The simplify names function converts district names to a more consistent format.
It does this by converting to lower case, removing spaces, removing "county" or "parish"
from the district name.
}
\details{
For example "New Haven County" and "new haven" would both be converted to "newhaven"

It does not correct for abbreviations. For example, it would not correctly
simplify Saint Lawrence, St. Lawrence and St Lawrence into a consistent name.

It does not differentiate between repeat county names from different states.
E.g. it would have no way to distinguish Jefferson County in Colorado from
Jefferson County in New York.

Written by J. Davis and M. Wimberly. Modified by A. Keyel. Original available at www.github.com/ecograph/ArboMAP. Used under GPL-3 license
}
